delete unused zod module
