+++ packages/@molt/command/src/CommandParameter/types.ts
@@ -1,7 +1,5 @@
 import type { Pam } from '../lib/Pam/index.js'
 
-export * from './zod.js'
-
 export type ArgumentValueMutuallyExclusive = {
   _tag: string
   value: Pam.Value
